Move OAuth client trait methods to internal trait
